requirements.txt
numpy==1.15.0
nltk==3.3.0
whoosh==2.7.4
pyyaml==3.13